Adds metaInfo protection logic to Apps client
