--- src/utils/apps/index.ts
@@ -1,2 +1,0 @@
-export * from './appFile'
-export * from './metaInfo'
