+++ src/utils/index.ts
@@ -5,4 +5,3 @@ export * from './time'
 export * from './billingOptions'
 export * from './domain'
 export * from './tenant'
-export * from './apps'
